All code functional
This was a large commit, many things were added, I became lazy trying to wrap
this assignment up as quickly as possible. It also appears that every line has
changed due to an earlier commit where all the indentation was changed to three
spaces. It is now back to four.